% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesPF.R
\name{bayesPF}
\alias{bayesPF}
\title{Fits a psychometric function}
\usage{
bayesPF(formula, data, link)
}
\arguments{
\item{formula}{A formula that one would pass to `glm` or similar.}

\item{data}{A data.frame or list}

\item{link}{A link function such as "logit" or "probit".}
}
\description{
Fits a psychometric function
}
