% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesPsychometric.R
\docType{package}
\name{BayesPsychometric}
\alias{BayesPsychometric}
\alias{BayesPsychometric-package}
\title{BayesPsychometric: A package for fitting psychometric functions using the power of Bayesian statistics and Hamiltonian Monte Carlo sampling.}
\description{
The BayesPsychometric provides one main function, bayesPF, and helper functions that operate on the output of bayesPF.
}
