% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_PSS_samples}
\alias{extract_PSS_samples}
\title{Calculate the posterior samples for the PSS}
\usage{
extract_PSS_samples(fit)
}
\value{
The posterior samples for the Point of Subjective Simultaneity
}
\description{
Calculate the posterior samples for the PSS
}
