% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractFactorSamples.R
\name{extractFactorSamples}
\alias{extractFactorSamples}
\title{Returns coefficient samples for mixed categories}
\usage{
extractFactorSamples(fit)
}
\arguments{
\item{fit}{An object returned from `bayesPF()`.}
}
\description{
Takes in a bayesPF fit object and returns the coefficient estimates for each
  combination of factors.
}
