% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f2stan.R
\name{f2stan}
\alias{f2stan}
\title{Convert a typical formula to a formula list that map2stan can understand}
\usage{
f2stan(formula, data, link = c("logit", "probit"),
  adaptive_pooling = FALSE)
}
\arguments{
\item{formula}{A formula that you would pass to `glm` or similar.}

\item{data}{A data frame of observations.}

\item{link}{A link function (either "logit" or "probit")}

\item{adaptive_pooling}{Logical (FALSE by default) Specifies whether
adaptive pooling should be used when fitting the model}
}
\value{
A model and data that can be used by `rstan::stan` as well as
  information about the model and the data.
}
\description{
Convert a typical formula to a formula list that map2stan can understand
}
