% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getLink}
\alias{getLink}
\title{Get the link function from a fitted model}
\usage{
getLink(fit)
}
\arguments{
\item{fit}{A model fit as returned by \code{\link{bayesPF}}}
}
\value{
The link and inverse link functions
}
\description{
Get the link function from a fitted model
}
