% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesPsychometric.R
\docType{package}
\name{BayesPsychometric}
\alias{BayesPsychometric}
\title{BayesPsychometric: A package for fitting psychometric functions using the
  power of Bayesian statistics and Hamiltonian Monte Carlo sampling.}
\description{
The BayesPsychometric provides two main functions: 1) `bayesPF` which will
  fit a logistic model using Stan, and 2) `f2stan` which builds up the
  Stan code from a formula and data frame.
}
