% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logit}
\alias{logit}
\title{Computes the log-odds of a number}
\usage{
logit(p)
}
\arguments{
\item{p}{A real number in the range \code{[0, 1]}}
}
\value{
The log-odds of \code{p}, a real number symmetric around 0
}
\description{
The function is defined by \deqn{\text{ln}\left(\frac{p}{1-p}\right)}{ln( p / ( 1 - p ) )}
This function has an inverse, the [logistic()] function.
}
